third_party/WebKit/LayoutTests/svg/custom/transform-ignore-after-invalid-expected.txt

CONSOLE ERROR: line 7: Error: <rect> attribute transform: Expected ')', "…e(2,2) scale(1,2,3) translate(50…".
